add followers and inbox screens, update user profile component, and enhance user fetching logic

+++ components/VideoComponent.tsx
@@ -11,7 +11,7 @@ export default function VideoComponent({ video, isViewable }: { video: any; isVi
         playerInstance.loop = true;
     });
 
-    const { user, likes, getLikes, followings } = useAuth();
+    const { user, likes, getLikes, followings, getFollowings } = useAuth();
     const router = useRouter();
     React.useEffect(() => {
         // Dynamically play or pause based on viewability
@@ -64,10 +64,11 @@ export default function VideoComponent({ video, isViewable }: { video: any; isVi
                 user_id: user?.id,
                 follower_user_id: video.User.id
             })
-        if (error) {
-            console.error(error);
+        console.log(data)
+        if (!error) {
+            getFollowings(user?.id);
         } else {
-            console.log(getFollowings(user?.id));
+            console.log(error)
         }
     };
 
@@ -75,7 +76,11 @@ export default function VideoComponent({ video, isViewable }: { video: any; isVi
         const { data, error } = await supabase.from('Follower')
             .delete()
             .eq('user_id', user?.id)
-            .eq('follower_id', video.User.id)
+            .eq('follower_user_id', video.User.id)
+        console.log(data)
+        if (!error) {
+            getFollowings(user?.id);
+        }
     };
     return (
         <View className='w-full h-full '>
@@ -95,14 +100,18 @@ export default function VideoComponent({ video, isViewable }: { video: any; isVi
                         <Text className='text-white text-2xl'>{video.User.username}</Text>
                     </View>
                     <View className='' >
-                        <View className=' items-center justify-between'>
+                        <View className=' items-center mb-2 justify-between'>
                             <TouchableOpacity  >
                                 <Ionicons name="person" size={32} color="white" onPress={() => router.push(`/user?user_id=${video.User.id}`)} />
                             </TouchableOpacity>
-                            <TouchableOpacity className=' absolute -bottom-1 left-4 bg-white mt-1 rounded-full justify-center items-center' >
-                                <Ionicons name="add" size={18} color="black" onPress={followerUser} />
-                            </TouchableOpacity>
-                            <Text className='absolute -bottom-5 left-3 text-white font-bold '>{followings?.filter((following: any) => following.follower_user_id === video.User.id).length}</Text>
+
+                            {followings?.filter((following: any) => following.follower_user_id === video.User.id).length > 0 ?
+                                (<TouchableOpacity className=' absolute -bottom-1 left-4 bg-white mt-1 rounded-full justify-center items-center' onPress={unfollowerUser} >
+                                    <Ionicons name="remove" size={18} color="black" />
+                                </TouchableOpacity>) : (<TouchableOpacity className=' absolute -bottom-1 left-4 bg-white mt-1 rounded-full justify-center items-center' onPress={followerUser} >
+                                    <Ionicons name="add" size={18} color="black" />
+                                </TouchableOpacity>)}
+                            <Text className='absolute -bottom-7 font-extrabold  left-3 text-white text-xl'>{followings?.length}</Text>
                         </View>
 
                         {likes?.filter((like: any) => like.video_id === video.id).length > 0 ? (
